+++ src/cn/ytxu/http_wrapper/config/property/api_data/ApiDataBean.java
@@ -1,6 +1,6 @@
 package cn.ytxu.http_wrapper.config.property.api_data;
 
-import cn.ytxu.http_wrapper.common.enums.ApiDataFileAddressType;
+import cn.ytxu.http_wrapper.common.enums.ApiDataFilePathType;
 import cn.ytxu.http_wrapper.common.enums.ApiDataSourceType;
 
 import java.util.Collections;
@@ -11,7 +11,7 @@ import java.util.List;
  */
 public class ApiDataBean {
     private String source = ApiDataSourceType.apidocjs.name();// API数据源的类型
-    private String file_address_type = ApiDataFileAddressType.file.name();// API数据文件的地址类型：现在有file与url两种
+    private String file_path_type = ApiDataFilePathType.file.name();// API数据文件的地址类型：现在有file与url两种
     private List<ApiDataFilePathInfoBean> file_path_infos = Collections.EMPTY_LIST;// 多操作系统的配置
     private String file_charset = "UTF-8";// API数据文件中内容的字符集；
 
@@ -19,8 +19,8 @@ public class ApiDataBean {
         return source;
     }
 
-    public String getFileAddressType() {
-        return file_address_type;
+    public String getFilePathType() {
+        return file_path_type;
     }
 
     public List<ApiDataFilePathInfoBean> getFilePathInfos() {